Reorganize a few tests
